optimized the size of EventItem; made it possible to use a long instead of String to represent memory address

+++ rv-predict-engine/src/main/java/rvpredict/log/LoggingEngine.java
@@ -111,7 +111,7 @@ public class LoggingEngine {
      * @param addrr additional information identifying the event
      * @param value data involved in the event
      */
-    public void saveEvent(EventType eventType, int id, long addrl, int addrr, long value) {
+    public void saveEvent(EventType eventType, int id, int addrl, int addrr, long value) {
         if (shutdown) return;
 
         long gid = globalEventID.incrementAndGet();
